Add files via upload
This is the original, it's pretty rough. I'm thinking someone might have improvements in wording or resources, or is able to create automation of filling in the blanks. Text in brackets is intedned to be customized. This is just as much a practiceo f how to use github as a sincere contribution in the vein of switching.software